chore: :lipstick: Global Style 폰트 적용 방법 변경
