+++ pages/_document.tsx
@@ -3,7 +3,13 @@ import { Head, Html, Main, NextScript } from 'next/document';
 function Document() {
   return (
     <Html>
-      <Head></Head>
+      <Head>
+        <link
+          rel="stylesheet"
+          as="style"
+          href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard-dynamic-subset.css"
+        />
+      </Head>
       <body>
         <Main />
         <NextScript />
